id is number of card and you can't add a card if a card with same number already exist

+++ back/server.js
@@ -9,7 +9,6 @@
 //import of needed express function
 const express = require("express");
 const bodyParser = require("body-parser");
-const uniqid = require("uniqid");
 
 //initialization of express app
 const app = express();
@@ -23,17 +22,29 @@ app.post("/api/updateData", (req, res) => {
   let fs = require("fs");
   let rawdata = fs.readFileSync("data.json");
   let data = JSON.parse(rawdata);
-  req.body.dataObject.idData = uniqid();
-  data.listeInfoBancaire.push(req.body.dataObject);
-  fs.writeFile(
-    "data.json",
-    JSON.stringify({ listeInfoBancaire: data.listeInfoBancaire }),
-    "utf8",
-    (err) => {
-      if (err) throw err;
-      console.log("Data written to file");
+  let cardAlreadyStorred = false;
+
+  data.listeInfoBancaire.forEach((element) => {
+    if (element.numeroCarte === req.body.dataObject.numeroCarte) {
+      cardAlreadyStorred = true;
     }
-  );
+  });
+
+  if (!cardAlreadyStorred) {
+    data.listeInfoBancaire.push(req.body.dataObject);
+    fs.writeFile(
+      "data.json",
+      JSON.stringify({ listeInfoBancaire: data.listeInfoBancaire }),
+      "utf8",
+      (err) => {
+        if (err) throw err;
+        console.log("Data written to file");
+        res.send("");
+      }
+    );
+  } else {
+    res.send("error");
+  }
 });
 
 //get to send the data to the front
